accelerators/pgr-ai-chatbot/requirements.txt
openai
redis
requests
python-dotenv
python-telegram-bot
PyYAML
pydub
ffmpeg

